Move store files to a directory
[MAILPOET-2987]
