+++ assets/js/src/form_editor/templates/store/actions.ts
