+++ assets/js/src/form_editor/templates/store/controls.ts
