+++ assets/js/src/form_editor/templates/store/reducer.ts
