+++ assets/js/src/form_editor/templates/store/store.ts
