+++ assets/js/src/form_editor/templates/store/types.ts
